requirements.txt
streamlit==1.45.1
plotly==5.22.0
pandas==2.3.0
numpy==2.1.2
matplotlib==3.10.3
folium==0.19.7
rasterio==1.4.3
pyproj==3.7.1
shapely==2.1.1
geopandas==1.1.0
odc==0.1.2
plotly==5.22.0
SpeechRecognition==3.10.0
fpdf2
qrcode[pil]
xgboost==1.7.5
earthengine-api
faiss-cpu
sentence-transformers
PyPDF2
whisper
pyttsx3
transformers
accelerate
safetensors
openai
